implementar service de produtos

+++ src/main/java/com/ecommerce/amazon/service/ProdutoService.java
@@ -1,77 +1,50 @@
 package com.ecommerce.amazon.service;
 
-import com.ecommerce.amazon.dto.produto.ProdutoRequestDTO;
-import com.ecommerce.amazon.dto.produto.ProdutoResponseDTO;
 import com.ecommerce.amazon.entity.Produto;
 import com.ecommerce.amazon.repository.ProdutoRepository;
-import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-@RequiredArgsConstructor
 public class ProdutoService {
 
-    private final ProdutoRepository produtoRepository;
+    @Autowired
+    ProdutoRepository produtoRepository;
 
-    public List<ProdutoResponseDTO> listarTodos() {
-        return produtoRepository.findAll()
-                .stream()
-                .map(this::converterParaDTO)
-                .toList();
+    public List<Produto> findAll() {
+        return produtoRepository.findAll();
     }
 
-    public ProdutoResponseDTO buscarPorId(Long id) {
-        Produto produto = produtoRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Produto não encontrado"));
-
-        return converterParaDTO(produto);
+    public Produto findById(Long id) {
+        return produtoRepository.findById(id).get();
     }
 
-    public ProdutoResponseDTO criar(ProdutoRequestDTO dto) {
-
-        Produto produto = Produto.builder()
-                .nome(dto.getNome())
-                .descricao(dto.getDescricao())
-                .preco(dto.getPreco())
-                .estoque(dto.getEstoque())
-                .ativo(dto.getAtivo())
-                .build();
-
-        produto = produtoRepository.save(produto);
-
-        return converterParaDTO(produto);
+    public Produto createProduto(Produto produto) {
+        return produtoRepository.save(produto);
     }
 
-    public ProdutoResponseDTO atualizar(Long id, ProdutoRequestDTO dto) {
+    public Produto updateProduto(Long id, Produto novoProduto) {
 
-        Produto produto = produtoRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Produto não encontrado"));
+        Produto produto = produtoRepository.findById(id).get();
 
-        produto.setNome(dto.getNome());
-        produto.setDescricao(dto.getDescricao());
-        produto.setPreco(dto.getPreco());
-        produto.setEstoque(dto.getEstoque());
-        produto.setAtivo(dto.getAtivo());
+        produto.setNome(novoProduto.getNome());
+        produto.setDescricao(novoProduto.getDescricao());
+        produto.setPreco(novoProduto.getPreco());
+        produto.setEstoque(novoProduto.getEstoque());
+        produto.setAtivo(novoProduto.getAtivo());
 
-        produto = produtoRepository.save(produto);
-
-        return converterParaDTO(produto);
+        return produtoRepository.save(produto);
     }
 
-    public void deletar(Long id) {
-        produtoRepository.deleteById(id);
-    }
+    public boolean deleteById(Long id) {
 
-    private ProdutoResponseDTO converterParaDTO(Produto produto) {
-        return ProdutoResponseDTO.builder()
-                .id(produto.getId())
-                .nome(produto.getNome())
-                .descricao(produto.getDescricao())
-                .preco(produto.getPreco())
-                .estoque(produto.getEstoque())
-                .ativo(produto.getAtivo())
-                .build();
+        if (produtoRepository.existsById(id)) {
+            produtoRepository.deleteById(id);
+            return true;
+        }
+
+        return false;
     }
 }
